src/requirements.txt
influxdb==5.3.0
requests==2.24.0
flask==1.1.2
python-dotenv==0.14.0
